% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{addPromoter}
\alias{addPromoter}
\title{add promoter to annotation}
\usage{
addPromoter(anno, PromoterLength = 1500, bedFile = NULL)
}
\arguments{
\item{anno}{anotation, imported gff/bed}

\item{PromoterLength}{the length of promoter region, default as 1500}

\item{bedFile}{the output bed file name}
}
\description{
add promoter to annotation
}
\examples{
data("geneHapR_test")
bed <- addPromoter(gff)
}
