% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{as.haplotype}
\alias{as.haplotype}
\title{as.haplotype}
\usage{
as.haplotype(hapResult)
}
\arguments{
\item{hapResult}{hapResult}
}
\value{
haplotype class
}
\description{
calculat Hap net from a hapSummary object
}
