% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessbigVCF.R
\name{bigVcfFilter}
\alias{bigVcfFilter}
\title{filterVCFin}
\usage{
bigVcfFilter(VCFin = "", VCFout = "",
              Chr = Chr, POS = POS, overWrite = TRUE)
}
\arguments{
\item{VCFin}{Path of input vcf file}

\item{VCFout}{Path of output vcf file}

\item{Chr}{CHROM name}

\item{POS}{a vector consist with start and end postion}

\item{overWrite}{Overwrite output file or not, defalt is TRUE}
}
\description{
filterVCFin
}
\examples{
\dontrun{
 bigVcfFilter(VCFin = "Ori.vcf.gz", VCFout = "filtered.vcf.gz",
              Chr = "scaffold_8", POS = c(19802,24501), overWrite = TRUE)
              }
}
