% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionalSites.R
\name{calcuSiteEffect}
\alias{calcuSiteEffect}
\title{Calculation of Sites Effective}
\usage{
calcuSiteEffect(
  hap,
  pheno,
  phenoNames = names(pheno),
  quality = FALSE,
  method = c("mrMLM", "FASTmrMLM", "FASTmrEMMA", "pLARmEB", "pKWmEB"),
  p.adj = "none"
)
}
\description{
Calculation of Sites Effective
}
