% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DataSet}
\alias{DataSet}
\alias{gff}
\alias{phenos}
\alias{vcf}
\alias{accGroup}
\title{Datasets
gff contains a sample of gff file used for test and samples}
\description{
Datasets
gff contains a sample of gff file used for test and samples

Datasets
phenos contains a simulated test pheno data used for test and samples

Datasets
vcf contains a vcf dataset used for test and samples

Datasets
data.frame defined the group of each accessions
}
\keyword{datasets}
