% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_table}
\alias{filter_table}
\title{filter variants stored in table}
\usage{
filter_table(
  x,
  mode = c("POS", "type", "both"),
  Chr = Chr,
  start = start,
  end = end,
  gff = gff,
  type = type,
  cusTyp = cusTyp,
  geneID = geneID
)
}
\arguments{
\item{x}{genotype dataset in hapmap format, object of data.frame class}

\item{mode}{filter mode, one of "POS", "type", "both"}

\item{Chr}{chromosome name, needed if mode set to "POS" or "both"}

\item{start}{start position, needed if mode set to "POS" or "both"}

\item{end}{end position, needed if mode set to "POS" or "both"}

\item{gff}{object of GRanges class, genome annotations imported by
\code{import_gff()}}

\item{type}{filter type, needed if mode set to "type" or "both",
one of "CDS", "exon", "gene", "genome", "custom",
if \code{type} was set to "custom", then \code{custom_type} is needed.}

\item{cusTyp}{character vector, custom filter type,
needed if \code{type} set to "custom"}

\item{geneID}{gene ID}
}
\description{
filter variants stored in table
}
\examples{
# example
data("geneHapR_test")
table <- filter_table(gt.geno, mode = "POS",
                      Chr = "scaffold_1", start = 4100, end = 5000)
}
