% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessbigVCF.R
\name{filterLargeVCF}
\alias{filterLargeVCF}
\title{Pre-process large vcf}
\usage{
filterLargeVCF(VCFin = "", VCFout = VCFout,
                Chr = Chr, POS = POS, override = TRUE)
}
\arguments{
\item{VCFin}{Path of input \verb{*.vcf/*.vcf.gz} file.}

\item{VCFout}{Path or path vector of output \verb{*.vcf/*.vcf.gz} file.}

\item{Chr}{a single \code{CHROM} name or \code{CHROM} names vector.}

\item{POS}{a vector consist with \code{start} and \code{end} position or a list
with length equal to \code{Chr}, each element consist with \code{start} and \code{end}
position.}

\item{override}{whether override existed file or not, default as \code{TRUE}.}
}
\description{
Filter/extract one or multiple gene(s)/range(s) from a large
\verb{*.vcf/*.vcf.gz} file.
}
\details{
This package import vcfs with \code{vcfR} which is more efficient to
import/manipulate vcf in R. However, import a large vcf file is time and
memory consuming. It's suggested that filter/extract variants in target
range with \code{filterLargeVCF}.

When filter/extract multi genes/ranges, the parameter of \code{Chr} and \code{POS}
must have equal length. Results will save to a single file if the user
provide a single file path or save to multiple vcf files when a equal length
vector consist with file paths is provided.

However, if you have hundreds gene/ranges need to extract from a very
large VCF, it's prefer to process with other linux tools in a script
on server such as: \code{vcftools} and \code{bcftools}.
}
\examples{
\dontrun{
 # extract a single gene/range from large vcf
 filterLargeVCF(VCFin = "Ori.vcf.gz", VCFout = "filtered.vcf.gz",
                Chr = "scaffold_8", POS = c(19802,24501), override = TRUE)

 # extract multi genes/ranges from large vcf
 filterLargeVCF(VCFin = "Ori.vcf.gz",
                VCFout = c("filtered1.vcf.gz",
                           "filtered2.vcf.gz",
                           "filtered3.vcf.gz"),
                Chr = c("scaffold_8",
                        "scaffold_8",
                        "scaffold_7"),
                POS = list(c(19802,24501),
                           c(27341,28949),
                           c(38469,40344)),
                override = TRUE)

}
}
