% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{geography}
\alias{geography}
\alias{plothapGeo}
\title{display geography distribution}
\usage{
plothapGeo(hap, AccINFO, Acc.col, LON.col, LAT.col, hapNames,
           mapRegion = "world", maplevel = "world",
           zColours = zColours,
           legend = "leftbottom", symbolSize = 1,
           barOrient='vert', maxZVal = 1, oceanCol="white",
           landCol="grey90", borderCol = "black", cex.legend = 0.8,
           ...)
}
\arguments{
\item{mapRegion}{a country name from getMap()[['NAME']] or
'world','africa','oceania','eurasia','uk' sets map extents, overrides
xlim,ylim}

\item{zColours}{colours to apply to the pie section for each attribute
column}

\item{symbolSize}{multiplier of default symbol size}

\item{maxZVal}{the attribute value corresponding to the maximum symbol size,
this can be used to set the scaling the same between multiple plots}

\item{oceanCol}{a colour for the ocean}

\item{landCol}{a colour to fill countries}

\item{borderCol}{the colour for country borders}

\item{...}{any extra arguments to points()}
}
\description{
display geography distribution
}
