% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2hap.R
\name{get_hap}
\alias{get_hap}
\title{generat haps from vcf}
\usage{
get_hap(vcf,
        hapPrefix = "H",
        filter_Chr = FALSE,
        Chr = "scaffold_1",
        filter_POS = FALSE,
        startPOS = 104,
        endPOS = 9889,
        hyb_remove = TRUE,
        na.drop = TRUE)
}
\arguments{
\item{vcf}{vcf}

\item{hapPrefix}{hapPrefix, defalt is "H"}

\item{filter_Chr}{filter vcf by Chrome or not, defalt is FALSE.
If TRUE, the Chr is needed.}

\item{Chr}{Needed for filter vcf by Chrom}

\item{filter_POS}{filter vcf by Position or not, defalt is FALSE.
If TRUE, startPOS and endPOS are needed.}

\item{startPOS, endPOS}{Needed for filter vcf by position.
startPOS must less than endPOS}

\item{hyb_remove}{Remove accessions contains hybrid site or not.
Defalt is TRUE.}

\item{na.drop}{Drop Accessions contains unknown allele site or note.
Defalt is TRUE}
}
\value{
data.frame, first four rows are fixed to meta information: CHR, POS, INFO, ALLELE
Hap names were placed in col1, Accessions were placed at the last col.
}
\description{
generate hap format from vcf.
}
\examples{

data("quickHap_test")
# You can import your vcf data by `import_vcf()` here
hap <- get_hap(vcf,hyb_remove = TRUE, na.drop =TRUE)
hapResult <- hap_result(hap)
plotHapTable(hapResult)
}
\author{
Zhangrenl
}
