% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenePOS.R
\name{getGenePOS}
\alias{getGenePOS}
\title{getGenePOS}
\usage{
getGenePOS(gff= gff,
           geneID = geneID,
           type = type,
           gffTermContaingeneID = "Parent")
}
\arguments{
\item{gff}{imported gff}

\item{geneID}{target geneID}

\item{type}{vector consist with one or more types in gff}

\item{gffTermContaingeneID}{which term contains the geneID in your gff,
defalt is Parent}
}
\value{
named vectors contains start, end and strand
}
\description{
getGenePOS
}
\examples{
data("quickHap_test")
genePOS <- getGenePOS(gff = gff,
           geneID = "test1G0387",
           type = "CDS",
           gffTermContaingeneID = "Parent")

}
