% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI_geneHapR.R
\name{GUI_geneHapR}
\alias{GUI_geneHapR}
\alias{build_app_geneHapR}
\alias{startGUI.geneHapR}
\title{the GUI interface of geneHapR}
\usage{
build_app_geneHapR(...)

startGUI.geneHapR()
}
\arguments{
\item{...}{not used}
}
\description{
the GUI interface of geneHapR

the GUI interface of geneHapR
}
\examples{
 app <- build_app_geneHapR()
 # startGUI.geneHapR()
}
