% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{hap2string}
\alias{hap2string}
\title{hap2DNAstring}
\usage{
hap2string(hapResult, type = "DNA")
}
\arguments{
\item{hapResult}{hapResult}

\item{type}{return a charactot vector or a DNAbin object}
}
\description{
convert hapsummary to BioStringsset class
}
