% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapVisualization.R
\name{hap_result}
\alias{hap_result}
\title{Summary hap results}
\usage{
hap_result(hap,
           hapPrefix = "H",
           file = file)
}
\arguments{
\item{hap}{an object of \code{haptypes} class, generated by \code{vcf2hap()} or \code{seqs2hap} or \code{import_hapResult()}}

\item{hapPrefix}{prefix of hap names, default as "H"}

\item{file}{file path where to save the hap summary result. If missing, nothing will save on disk.}
}
\value{
data.frame, first four rows are fixed to meta information: CHR, POS, INFO, ALLELE
Hap names were placed in col1, Accessions and freqs were placed at the last two cols.
}
\description{
A function summarize hap result for visualization and calculation.
}
\details{
It is suggested that use the result of \code{vcf2hap()} or \code{seqs2hap} as input directly.
But user can read hap result from file by \code{importHapResult()}
}
\note{
Since a multi-letter prefix of hap names is possible, \code{hapPrefix} is needed
if the user have changed the default \code{hapPrefix} in \code{vcf2hap()} or \code{seqs2hap}.
}
\examples{
\dontrun{
# load example data
data("quickHap_test")

# generate hap result from vcfR object
hap <- vcf2hap(vcf, hyb_remove = TRUE, na.drop =TRUE)

# generate hap result from DNAStringSet object
hap <- seqs2hap(seqs, Ref = names(seqs)[1], needAlign = TRUE, needTrim = TRUE)

# summary hap result for plot hap table
hapResult <- hap_result(hap)

# plot hap table
plotHapTable(hapResult)

# user could change/add themes, title, ...
p <- plotHapTable(hapResult)
p + ggplot2::theme_classic()
p + ggplot2::labs(title = "new title")
}
}
