% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setATGas0.R
\name{hapSetATGas0}
\alias{hapSetATGas0}
\title{hapSetATGas0}
\usage{
hapSetATGas0(gff = gff, hap = hap,
             geneID = geneID,
             Chr = Chr, POS = c(start, end))
}
\arguments{
\item{gff}{imported gff}

\item{hap}{hap result with meta infos}

\item{geneID}{geneID}

\item{Chr}{Chrom name}

\item{POS}{vector defined by start and end position}
}
\value{
hap results with set ATG position as 0
}
\description{
set ATG posision as 0
}
\examples{
data("quickHap_test")
hap <- get_hap(vcf)
hapn <- hapSetATGas0(gff = gff, hap = hap,
                     geneID = "test1G0387",
                     Chr = "scaffold_1",
                     POS = c(4300, 7910))
}
