% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_phenos.R
\name{hapVsPhenoPerSite}
\alias{hapVsPhenoPerSite}
\title{hapVsPhenoPerSite}
\usage{
hapVsPhenoPerSite(
  hap,
  pheno,
  phenoName,
  sitePOS,
  fileName,
  fileType = NULL,
  freq.min = 5,
  ...
)
}
\arguments{
\item{hap}{an R object of hapresult class}

\item{pheno, phenoName}{pheno, a data.frame contains the phenotypes;
Only one phenotype name is required.}

\item{sitePOS}{the coordinate of site}

\item{fileName, fileType}{file name and file type will be needed for saving result,
file type could be one of "png, tiff, jpg"}

\item{freq.min}{miner allies frequency less than freq.min will not be skipped}

\item{...}{addtional params will be passed to plot saving function like \code{tiff()}, \code{png()}, \code{pdf()}}
}
\description{
Comparie phenotype site by site.
}
\examples{
data("geneHapR_test")
hapVsPhenoPerSite(hapResult, pheno, sitePOS = "4300")
}
