% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenos.R
\name{hapVsPhenos}
\alias{hapVsPhenos}
\title{hapVsPhenoS}
\usage{
hapVsPhenos(hap,
            phenos,
            outPutSingleFile = TRUE,
            hapPrefix = "H",
            geneID = "Seita.0G000000",
            mergeFigs = TRUE,
            file = file,
            width = 12,
            height = 8, ...)
}
\arguments{
\item{hap}{hap}

\item{phenos}{phenos}

\item{outPutSingleFile}{only worked while file type is pdf.
Out all figs in a single pdf file, figs will plot in each pages.}

\item{hapPrefix}{hap  prefix}

\item{geneID}{gene ID forplot title}

\item{mergeFigs}{merge heatmapand box plot or not}

\item{file}{out put file name. File type could be pdf, png, tiff, jpg, bmp.}

\item{width}{manual option for determining the output file width in inches.
(defalt: 12)}

\item{height}{manual option for determining the output file height in inches.
(defalt: 8)}

\item{...}{options will pass to ggpubr}
}
\description{
hapVsPhenoS
}
