% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_hapResult}
\alias{import_hapResult}
\title{import_hapResult}
\usage{
import_hapResult(file)
}
\arguments{
\item{file}{hapResult file path}
}
\value{
data.frame
}
\description{
import_hapResult
}
\examples{
\dontrun{
    hapResult <- import_hapResult("your_hapResult_file.txt")
}
}
