% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_pheno}
\alias{import_pheno}
\title{imports phenos from file}
\usage{
import_pheno(phenoFile, comment.char = "#", ...)
}
\arguments{
\item{phenoFile}{pheno file path, should be a table separated by tab}

\item{comment.char}{comment.char, start with comment.char will be ignored}

\item{...}{parameters will pass to read.delim}
}
\value{
data.frame, Accession names were set as rownames and cols were
named by pheno names
}
\description{
first col should be Accessions
phenos should loacted begin second col，
phenoName should at the first row,
If "." located in pheno name, the former part of phenoName will be set as y axis name
and the latter part will be set as foot of the fig.
}
\examples{
\dontrun{

pheno <- import_pheno("test.pheno.txt")
}
}
