% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_seqs}
\alias{import_seqs}
\title{import_seqs}
\usage{
import_seqs(file, format = "fasta", ...)
}
\arguments{
\item{file}{gff file path}

\item{format}{Either "fasta" (the default) or "fastq"}

\item{...}{Others parameters supported by Biostrings::readDNAStringSet}
}
\value{
DNAstringSet
}
\description{
import_seqs
}
\examples{
\dontrun{
   geneSeqs <- import_seqs(file = "fastaFilePath", format = "fasta")
}
}
