% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network}
\alias{network}
\alias{get_hapNet}
\title{generate haplotype net relationshop with haplotype result}
\usage{
get_hapNet(hapResult,
           accGroup = accGroup,
           groupName = groupName)
}
\arguments{
\item{hapResult}{object of \code{hapSummary} class, generated by \code{hap_result()}}

\item{accGroup}{data.frame, specified groups of each accession.
Used for pie plot. If missing, pie will not draw in plotHapNet.
Or you can supplied a hap_group mattrix with \code{plot(hapNet, pie = hap_group)}.}

\item{groupName}{the group name used for pie plot,
should be in \code{accGroup} column names, default as the first column name}
}
\value{
haplonet class
}
\description{
computes a haplotype network with haplotype summary result
}
\examples{

# load sample datasets
data("quickHap_test")

# generate haplotype result from vcfR
hap <- vcf2hap(vcf)
hapResult <- hap_result(hap)

# calculate haploNet
hapNet <- get_hapNet(hapResult,
                     accGroup = accGroup, # accession types
                     groupName = colnames(accGroup)[2])

# plot haploNet
plot(hapNet)
# plot haploNet
plotHapNet(hapNet,
           size = "freq", scale.ratio = 1,  # circle size
           scale = TRUE, # scale circle with 'log10(size + 1)'
           cex = 1, # size of hap symbol
           col.link = 2, # link colors
           link.width = 2, # link widths
           show.mutation = 2, # mutation types one of c(0,1,2,3)
           legendPosition = "right") # legend position
}
\references{
Article{,
title = {The stringdist package for approximate string matching},
author = {M.P.J. {van der Loo}},
year = {2014},
journal = {The {R} {J}ournal},
volume = {6},
issue = {1},
url = {https://CRAN.R-project.org/package=stringdist},
pages = {111-122},
}
Article{,
title = {pegas: an {R} package for population genetics with an integrated--modular approach},
author = {E. Paradis},
journal = {Bioinformatics},
year = {2010},
volume = {26},
pages = {419-420},
}
}
\seealso{
\code{\link[geneHapR:plotHapNet]{plotHapNet()}} and
\code{\link[geneHapR:hap_result]{hap_result()}}.
}
