% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_gridhaptable2.R
\name{plotHapTable2}
\alias{plotHapTable2}
\title{plotHapTable2}
\usage{
plotHapTable2(
  hapSummary,
  show_indel_size = 1,
  replaceMultiAllele = TRUE,
  show_INFO = FALSE,
  title = "",
  gff = NULL,
  show_chr_name = TRUE,
  Chr = NULL,
  start = NULL,
  end = NULL,
  gene_model_fill = "white",
  gene_model_height = 0.2,
  space_height = 0.1,
  table_height = NULL,
  CDS_height = 0.1,
  headrows = 1,
  equal_col_width = FALSE,
  head_anno = 1,
  col_annots = 0,
  row_labels = 1,
  row_annots = 1,
  labels_fill = "white",
  annot_fill = "grey90",
  head_fill = NULL,
  cell_fill = NULL,
  style = gpar(fontfamily = "sans", fontface = 1, cex = 0.7),
  footbar = ""
)
}
\arguments{
\item{hapSummary}{the hapSummary or hapResult object}

\item{show_indel_size}{the Indel length longer will be replaced by "i1,i2,i3,..."}

\item{replaceMultiAllele}{replace multi-allele title by 'T1,T2,...' or not}

\item{show_INFO}{show annotation field or not, default as \code{FALSE}}

\item{title}{title of plot}

\item{gff}{gff or bed annotation}

\item{show_chr_name}{show chromosome name at left-top cell or not}

\item{Chr, start, end}{which range should be plotted in gene model}

\item{gene_model_fill}{fill color of CDS ad UTR}

\item{gene_model_height, table_height, space_height}{the plotting range height of gene model, table and spacer}

\item{CDS_height}{a numeric vector specified the height of CDS, and the height of utr is half of that,
only useful when gff is provided,}

\item{equal_col_width}{a bool or numeric vector specified whether column with should be equal}

\item{col_annots, head_anno, headrows, row_annots, row_labels}{the column or row number of annotation or labels or heads}

\item{annot_fill, head_fill, labels_fill}{the fill color of annotation, head and label row or columns}

\item{cell_fill}{a color vector or function or named vector specified cell fill color}

\item{footbar}{the foot notes}
}
\description{
plot the hapResult in table like style using grid system.
}
\examples{
#
data(geneHapR_test)
plotHapTable2(hapResult)
plotHapTable2(hapResult, gff = gff)
}
