% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta2hap.R
\name{seqs2hap}
\alias{seqs2hap}
\title{generate hap results from seqs}
\usage{
seqs2hap(seqs,
         Ref = names(seqs)[1],
         hyb_remove = TRUE, na.drop = TRUE,
         needAlign = TRUE, needTrim = TRUE,
         minFlankFraction = 0.05, maxGapsPerSeq = 0.2,
         hapPrefix = "H", ...)
}
\arguments{
\item{seqs}{DNAStringSet or DNAMultipleAlignment}

\item{Ref}{T}

\item{hyb_remove}{T}

\item{na.drop}{T}

\item{needAlign}{T}

\item{needTrim}{T}

\item{minFlankFraction}{A value in \verb{[0, 1]} that indicates the minimum fraction needed to call a gap in the consensus string (default is 0.1).}

\item{maxGapsPerSeq}{A value in \verb{[0, 1]} that indicates the maximum gaps allowed in a seq after alignment}

\item{hapPrefix}{T}

\item{...}{parameters will pass to musle::musle() for DNA multi alignment}
}
\description{
generate hap results from seqs
}
\examples{
\dontrun{
# load example data
data("quickHap_test")

# generate hap result from vcfR object
hap <- vcf2hap(vcf, hyb_remove = TRUE, na.drop =TRUE)

# generate hap result from DNAStringSet object
hap <- seqs2hap(seqs, Ref = names(seqs)[1], needAlign = TRUE, needTrim = TRUE)

# summary hap result for plot hap table
hapResult <- hap_result(hap)

# plot hap table
plotHapTable(hapResult)

# user could change/add themes, title, ...
p <- plotHapTable(hapResult)
p + ggplot2::theme_classic()
p + ggplot2::labs(title = "new title")
}
}
