% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionalSites.R
\name{siteFunEff}
\alias{siteFunEff}
\alias{plotSiteFunEff}
\title{Find site functional effective}
\usage{
siteFunEff(hap, pheno, phenoNames)

plotSiteFunEff(results,
               title = title,
               caption = caption,
               facet = FALSE,
               ...)
}
\arguments{
\item{hap}{object of "hapResult" class}

\item{pheno}{phenotype data, with column names as pheno name
and row name as accessions.}

\item{phenoNames}{pheno names used for analysis, if missing,
will use all pheno names in \code{pheno}}

\item{results}{results of site functional effect analysis}

\item{title}{The text for the title.}

\item{caption}{The text for the caption which will be displayed in the
bottom-right of the plot by default.}

\item{facet}{use facet or not, default as \code{FALSE}}

\item{...}{A list of new name-value pairs. The name should be an aesthetic.}
}
\value{
data.frame, with column name as pheno name and
row name as site position
}
\description{
Find site functional effective
}
\examples{
\donttest{
data("geneHapR_test")

# calculate site functional effect
funResults <- siteFunEff(hapResult, pheno, names(pheno))
plotSiteFunEff(funResults, facet = FALSE)
}
}
