% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_functionalSites.R
\name{sites_compar}
\alias{sites_compar}
\alias{compareAllSites}
\title{sites comparison}
\usage{
compareAllSites(
  hap,
  pheno,
  phenoName = names(pheno)[1],
  hetero_remove = TRUE,
  title = "",
  file = file
)
}
\arguments{
\item{hap}{object of hapResult class}

\item{pheno}{a data.frame contains phenotypes}

\item{phenoName}{the name of used phenotype}

\item{hetero_remove}{removing the heter-sites or not, default as TRUE}

\item{title}{the title of the figure}

\item{file}{if provieds a file path the comparing results will saved to file.}
}
\description{
Used for all allele effect compare once
}
