% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2hap.R
\name{vcf2hap_data}
\alias{vcf2hap_data}
\title{vcf2hap_data}
\usage{
vcf2hap_data(
  vcf,
  allS_new = allS_new,
  REF = REF,
  ALT = ALT,
  ALLELE = ALLELE,
  POS = POS
)
}
\value{
data.frame, individuals in rows and positions in cols
with additional attrs
eg. # $hap
#    41  136
# a  A   A|T
# b  G   G
# $allS_new
}
\description{
vcf2hap_data
}
