% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{writeout}
\alias{writeout}
\title{writeout}
\usage{
writeout(x, file, ...)
}
\arguments{
\item{x}{objec of haplotypes or hapSummary}

\item{file}{filepath}

\item{...}{other parameters will be drop}
}
\description{
writeout
}
