% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addINFO.R
\name{addINFO}
\alias{addINFO}
\title{Summary hap results}
\usage{
addINFO(hapResult,
        tag = "", values = values,
        replace = FALSE, sep = ";")
}
\arguments{
\item{hapResult}{object of \code{haptypes} class, generated by \code{vcf2hap()}
or \code{seqs2hap} or \code{import_hapResult()}}

\item{tag}{tag names. Usually is a single word used before "="}

\item{values}{annotation for each site.
Length of values must be equal with sites in hap result}

\item{replace}{whether replace origin INFOs in hap result or not.
Default as FALSE}

\item{sep}{a character string to separate the terms.  Not
    \code{\link[base]{NA_character_}}.}
}
\description{
add annotations to \code{INFO} fields used for \code{plotHapTable()}
}
\examples{
\dontrun{
data("quickHap_test")
hap <- vcf2hap(vcf)

# length of values must be equal with number of sites in hap result
values <- paste0("newInfo",c(1:9))
hap <- addINFO(hap, tag = "new", values = values, replace = TRUE)
}
}
\seealso{
\code{\link[geneHapR:plotHapTable]{plotHapTable()}}
}
