% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{as.haplotype}
\alias{as.haplotype}
\title{as.haplotype}
\usage{
as.haplotype(hapResult)
}
\arguments{
\item{hapResult}{hapResult}
}
\value{
haplotype class
}
\description{
convert hapSummary into haplotype (pegas)
}
\note{
It's not advised for hapSummary with indels, due to indels will replaced by equal length of SNPs.
}
