% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{ashaplotype}
\alias{ashaplotype}
\alias{as.haplotype}
\title{as.haplotype}
\usage{
as.haplotype(hapResult)
}
\arguments{
\item{hapResult}{object of \code{hapSummary} class, generated by \code{hap_result()}}
}
\value{
haplotype class
}
\description{
convert hapSummary into haplotype (pegas)
}
\note{
It's not advised for hapSummary with indels, due to indels will
convert to SNPs with equal length of each indel.
}
