% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DataSet}
\alias{DataSet}
\alias{gff}
\alias{seqs}
\alias{pheno}
\alias{vcf}
\alias{accGroup}
\title{Datasets
\code{gff} contains a example of gff file used for test of
visualization mutations on gene model.}
\description{
\code{pheno} contains a simulated test pheno data used for test of comparison
between different haps

\code{vcf}, a vcfR object provide a data set for test of \code{seq2hap()}.
\code{vcf} contains indels, snps, biallelic sites and multiallelic sites.

\code{accGroup} a data.frame defined the group of each accessions
}
\keyword{datasets}
