% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtor.R
\name{filter_hap}
\alias{filter_hap}
\title{filter_hap_by_gff}
\usage{
filter_hap(gff = gff,
           hap = hap,
           hap_prefix_new = "C_H",
           type = c("CDS", "exon", "gene","genome", "custom"),
           custom_type = c("CDS", "exon", "gene"))
}
\arguments{
\item{gff}{gff文件}

\item{hap}{hap或hapResult}

\item{hap_prefix_new}{new prefix for filtered haps}

\item{type}{筛选模式，CDS/exon/gene/genome/custom之一，
如果是custom则必须设定custom_type}

\item{custom_type}{自定义筛选模式}
}
\description{
根据变异位点所在区域的类型进行筛选
}
\examples{
# filtet hap
data("quickHap_test")
hap <- get_hap(vcf)
hap <- filter_hap(gff = gff, hap = hap, type = "CDS")

}
