% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtor.R
\name{filter_vcf}
\alias{filter_vcf}
\title{filter_vcf_by_gff}
\usage{
filter_vcf(vcf, gff = gff,
           mode = c("POS", "type", "both"),
           Chr = Chr, start = start, end = end,
           type = c("CDS", "exon", "gene", "genome", "custom"),
           cusTyp = c("CDS", "five_prime_UTR", "three_prime_UTR"))
}
\arguments{
\item{vcf}{imported vcf}

\item{gff}{imported gff}

\item{mode}{filter mode, one of POS/type/both}

\item{Chr}{CHROM name, needed if mode set to 'POS' or 'both'}

\item{start}{Start position, needed if mode set to 'POS' or 'both'}

\item{end}{End position, needed if mode set to 'POS' or 'both'}

\item{type}{filter type, needed if mode set to 'type' or 'both',
one of CDS/exon/gene/genome/custom,
if type set to custom, the custom_type is needed.}

\item{cusTyp}{vector, custom filter type, needed if type set to custom}
}
\description{
filter vcf by gff file or by position or both
}
\examples{
# filtet hap
data("quickHap_test")
vcf_f1 <- filter_vcf(vcf, mode = "POS",
                    Chr = "scaffold_1",
                    start = 4300, end = 5890)

vcf_f2 <- filter_vcf(vcf, mode = "type",
                    gff = gff,
                    type = "CDS")
vcf_f3 <- filter_vcf(vcf, mode = "both",
                    Chr = "scaffold_1",
                    start = 4300, end = 5890,
                    gff = gff,
                    type = "CDS")

}
