% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneHapR-package.R
\docType{package}
\name{geneHapR-package}
\alias{geneHapR}
\alias{geneHapR-package}
\title{geneHapR: R Packages for Gene Haplotype Analysis}
\description{
This packages provides a suite for one or multi-genes haplotype analysis,
visualizion and phenotypes analysis.
}
\details{
The haplotype analysis requires sequences such as fasta which could be
import by \link[geneHapR:import_vcf]{import_vcf}

\strong{A complete list of functions} can be displayed with:
\code{library(help = geneHapR)}.

\strong{Vignettes} (documentation) can be listed with:
\code{browseVignettes('geneHapR')}.

Several small example \strong{datasets} are included in geneHapR.
\strong{vcf} and \strong{seqs} provide a data set for test of \code{seq2hap()} and
\code{seqs2hap()}.
\strong{gff} provide a data set for visualization mutations on gene model.
\strong{pheno} provide a data set for phenotype comparison between haplotypes.
\strong{accGroup} provide a data set for pie plot in hapNet visualization.
}
\seealso{
More documentation for vcfR can be found at repos stored at
\href{https://gitee.com/zhangrenl/genehapr}{gitee} and
\href{https://github.com/zhangrenl/genehapr}{github}
}
\author{
\strong{Maintainer}: Zhang Renliang \email{zhang_renliang@163.com}

Authors:
\itemize{
  \item Jia Guanqing \email{jiaguanqing@caas.cn}
}

}
\keyword{internal}
