% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{get_hapNet}
\alias{get_hapNet}
\title{generate haplotype net relationshop with haplotype result}
\usage{
get_hapNet(hapResult, accGroup = accGroup, groupName = groupName)
}
\arguments{
\item{hapResult}{hapResult}

\item{accGroup}{data.frame, specified groups of each accession.
Used for pie plot. If missing, pie will not draw in plotHapNet.
Or you can supplied a hap.group mattrix with plot(hapNet, pie = hap.group).}

\item{groupName}{the group name used for pie plot, should be one of accGroup colnames, defalt as the first column name}
}
\value{
haplonet class
}
\description{
get_hapNet computes a haplotype network with haplotype summary result.
}
\references{
Article{,
title = {The stringdist package for approximate string matching},
author = {M.P.J. {van der Loo}},
year = {2014},
journal = {The {R} {J}ournal},
volume = {6},
issue = {1},
url = {https://CRAN.R-project.org/package=stringdist},
pages = {111-122},
}
Article{,
title = {pegas: an {R} package for population genetics with an integrated--modular approach},
author = {E. Paradis},
journal = {Bioinformatics},
year = {2010},
volume = {26},
pages = {419-420},
}
}
\seealso{
plot, plothapNet, hap_result
}
