% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenos.R
\name{hapVsPhenos}
\alias{hapVsPhenos}
\title{hapVsPhenos}
\usage{
hapVsPhenos(hap,
            pheno,
            outPutSingleFile = TRUE,
            hapPrefix = "H",
            title = "Seita.0G000000",
            file = file,
            width = 12,
            height = 8,
            res = res, ...)
}
\arguments{
\item{outPutSingleFile}{\code{TRUE} or \code{FALSE} indicate whether put all figs
into to each pages of single file or generate multi-files.
Only worked while file type is pdf}

\item{file}{the output file name/path. File type could be pdf, png, tiff,
jpg or bmp. \code{file} will only used for out file name/path directly if file
type is "pdf" and \code{outPutSingleFile} set as \code{TRUE}. Otherwise the pheno names
will added for each file name/path.}

\item{width}{manual option for determining the output file width in inches.
(default: 12)}

\item{height}{manual option for determining the output file height in inches.
(default: 8)}

\item{res}{The nominal resolution in ppi which will be recorded in the
bitmap file, if a positive integer. Also used for units other than the
default, and to convert points to pixels.}
}
\description{
hapVsPhenos
}
\examples{

\dontrun{
# analysis all pheno in the data.frame of pheno
hapVsPhenos(hap,
            pheno,
            outPutSingleFile = TRUE,
            hapPrefix = "H",
            title = "Seita.0G000000",
            file = "mypheno.tiff",
            width = 12,
            height = 8,
            res = 300)
}
}
