% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{import_AccINFO}
\alias{import_AccINFO}
\title{imports accession information from file}
\usage{
import_AccINFO(file, comment.char = "#",
               check.names = FALSE, row.names = 1, ...)
}
\arguments{
\item{file}{file path, this file should be a tab delimed table}

\item{comment.char}{comment.char, start with comment.char will be ignored}

\item{check.names}{logical.  If \code{TRUE} then the names of the
    variables in the data frame are checked to ensure that they are
    syntactically valid variable names.  If necessary they are adjusted
    (by \code{\link{make.names}}) so that they are, and also to ensure
    that there are no duplicates.}

\item{row.names}{a vector of row names.  This can be a vector giving
    the actual row names, or a single number giving the column of the
    table which contains the row names, or character string giving the
    name of the table column containing the row names.

    If there is a header and the first row contains one fewer field than
    the number of columns, the first column in the input is used for the
    row names.  Otherwise if \code{row.names} is missing, the rows are
    numbered.

    Using \code{row.names = NULL} forces row numbering. Missing or
    \code{NULL} \code{row.names} generate row names that are considered
    to be \sQuote{automatic} (and not preserved by \code{\link{as.matrix}}).
  }

\item{...}{parameters will pass to read.delim}
}
\value{
data.frame, Accession names were set as rownames and cols were
named by pheno names
}
\description{
import accession information including phenotype data,
accession group, location from a tab delimed table file
}
\details{
First column should be Accessions;
phenos should begin from second col,
phenoName/group/locations should located at the first row,
If a dot '.' is located in pheno name, then
the part before the dot will be set as y axis name
while the followed will be set as foot of the fig.
}
\examples{
\dontrun{

pheno <- import_AccINFO("test.pheno.txt")
}
}
