% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{import_gff}
\alias{import_gff}
\title{Import Annotations from GFF Format File}
\usage{
import_gff(gffFile, format = "GFF")
}
\arguments{
\item{gffFile}{the gff file path}

\item{format}{should be one of "gff", "gff1", "gff2", "gff3", "gvf",
or "gtf". Default as GFF}
}
\value{
GRange object
}
\description{
import genome annotations in GFF/GFF3 format
}
\examples{
\dontrun{

    gff <- import_gff("your.gff", format = "GFF")
}
}
