% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{import_hap}
\alias{import_hap}
\title{import_hap}
\usage{
import_hap(file, ...)
}
\arguments{
\item{file}{hapSummary or hapResult file path}

\item{...}{extras will pass to \code{read.delim()}}
}
\value{
hapSummary or hapResult
}
\description{
This function could be used for import hap result or hap summary result.
The type of returned object is decided by hap result format, see details.
}
\details{
The hap result and hap summary result have common features.
The common features of these two types are:
First four rows contains extra information: CHR, POS, INFO and ALLELE
Hap names were in the first column.
The differences are:
Hap summary result have a freq column while hap result not.
Rows represent haplotypes in hap summary result, while rows represent accessions in hap result.
In addtion, the accessions of each haplotype in hap summary result were separated by ';'.
}
\examples{
\dontrun{

hap <- import_hap("hapSummary_OR_hapResult_file.txt")
}
}
