% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_MultipleAlignment}
\alias{import_MultipleAlignment}
\title{import MultipleAlignment}
\usage{
import_MultipleAlignment(filepath, format = "fasta", type = "DNA")
}
\arguments{
\item{filepath}{
    A character vector (of arbitrary length when reading, of length 1
    when writing) containing the paths to the files to read or write.
    Note that special values like \code{""} or \code{"|cmd"} (typically
    supported by other I/O functions in R) are not supported here. Also
    \code{filepath} cannot be a connection.
  }

\item{format}{
    Either \code{"fasta"} (the default), \code{stockholm}, or
    \code{"clustal"}.
  }

\item{type}{one of 'DNA' and 'Protein'}
}
\description{
import MultipleAlignment
}
\examples{
\dontrun{
   geneSeqs <- import_MultipleAlignment(filepath = "fastaFilePath",
                                        format = "fasta",
                                         type = "DNA")
   geneSeqs <- import_MultipleAlignment(filepath = "fastaFilePath",
                                        format = "fasta",
                                         type = "Protein")
}
}
