% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_pheno}
\alias{import_pheno}
\title{imports phenos from file}
\usage{
import_pheno(phenoFile, comment.char = "#", ...)
}
\arguments{
\item{phenoFile}{pheno file path, should be a table separated by tab}

\item{comment.char}{comment.char, start with comment.char will be ignored}

\item{...}{parameters will pass to read.delim}
}
\value{
data.frame, Accession names were set as rownames and cols were
named by pheno names
}
\description{
import phenos from table format file
}
\details{
First column should be Accessions;
phenos should begin from second col，
phenoName should located at the first row,
If a dot '.' is located in pheno name, then
the part before the dot will be set as y axis name
while the followed will be set as foot of the fig.
}
\examples{
\dontrun{

pheno <- import_pheno("test.pheno.txt")
}
}
