% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{import_seqs}
\alias{import_seqs}
\title{import_seqs}
\usage{
import_seqs(filepath, format = "fasta")
}
\arguments{
\item{filepath}{A character vector containing the path(s) to the file(s)
to read or write.
Reading files in gzip format (which usually have the '.gz' extension) is
supported.
\emph{Note} that only DNA supported here.}

\item{format}{Either \code{"fasta"} (the default) or \code{"fastq"}}
}
\description{
import_seqs
}
\examples{
\dontrun{
   geneSeqs <- import_seqs(filepath = "fastaFilePath", format = "fasta")
}
}
