% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_vcf}
\alias{import_vcf}
\title{import vcf from file}
\usage{
import_vcf(vcf_file = vcf_file, ...)

import_vcf(vcf_file = vcf_file, ...)
}
\arguments{
\item{vcf_file}{file path of vcf}

\item{...}{pass to vcfR::read.vcfR}
}
\value{
vcfR object
}
\description{
Read and files in the \verb{*.vcf} structured text format,
as well as the compressed \verb{*.vcf.gz} format.
}
\examples{
\dontrun{

vcf <- import_vcf(file = "test.vcf.gz")
}
}
\seealso{
\code{\link[vcfR:read.vcfR]{vcfR::read.vcfR()}}
}
\author{
Zhangrenl
}
