% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{import_vcf}
\alias{import_vcf}
\title{Import VCF from File}
\usage{
import_vcf(vcf_file = vcf_file, ...)

import_vcf(vcf_file = vcf_file, ...)
}
\arguments{
\item{vcf_file}{file path of VCF file}

\item{...}{pass to \code{vcfR::read.vcfR()}}
}
\value{
vcfR object
}
\description{
import *.vcf structured text format,
as well as the compressed \verb{*.vcf.gz} format.
}
\examples{

vcf.Path <- system.file("extdata", "var.vcf.gz", package = "geneHapR")
vcf <- import_vcf(file = vcf.Path)
vcf

}
\seealso{
\code{\link[vcfR:read.vcfR]{vcfR::read.vcfR()}}
}
\author{
Zhangrenl
}
