% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network}
\alias{network}
\alias{get_hapNet}
\title{generate haplotype net relationshop with haplotype result}
\usage{
get_hapNet(hapSummary,
           AccINFO = AccINFO,
           groupName = groupName)
}
\arguments{
\item{hapSummary}{object of \code{hapSummary} class, generated by \code{hap_summary()}}

\item{AccINFO}{data.frame, specified groups of each accession.
Used for pie plot. If missing, pie will not draw in plotHapNet.
Or you can supplied a hap_group mattrix with \code{plot(hapNet, pie = hap_group)}.}

\item{groupName}{the group name used for pie plot,
should be in \code{AccINFO} column names, default as the first column name}
}
\value{
haplonet class
}
\description{
computes a haplotype network with haplotype summary result
}
\examples{

\donttest{
# load sample datasets
data("geneHapR_test")
hapSummary <- hap_summary(hapResult)

# calculate haploNet
hapNet <- get_hapNet(hapSummary,
                     AccINFO = AccINFO, # accession types
                     groupName = colnames(AccINFO)[2])

# plot haploNet
plot(hapNet)

# plot haploNet
plotHapNet(hapNet,
           size = "freq",   # circle size
           scale = "log10", # scale circle with 'log10(size + 1)'
           cex = 1, # size of hap symbol
           col.link = 2, # link colors
           link.width = 2, # link widths
           show.mutation = 2, # mutation types one of c(0,1,2,3)
           legend = FALSE) # legend position
}
}
\references{
Article{,
title = {The stringdist package for approximate string matching},
author = {M.P.J. {van der Loo}},
year = {2014},
journal = {The {R} {J}ournal},
volume = {6},
issue = {1},
url = {https://CRAN.R-project.org/package=stringdist},
pages = {111-122},
}
Article{,
title = {pegas: an {R} package for population genetics with an integrated--modular approach},
author = {E. Paradis},
journal = {Bioinformatics},
year = {2010},
volume = {26},
pages = {419-420},
}
}
\seealso{
\code{\link[geneHapR:plotHapNet]{plotHapNet()}} and
\code{\link[geneHapR:hap_summary]{hap_summary()}}.
}
