% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{plotHapNet}
\alias{plotHapNet}
\title{plotHapNet}
\usage{
plotHapNet(
  hapNet,
  size = "freq",
  scale.ratio = 1,
  cex = 0.8,
  col.link = 1,
  link.width = link.width,
  show.mutation = 1,
  lwd = 1,
  pieCol = pieCol,
  pie = hapGroup,
  legendPOS = "left",
  ...
)
}
\arguments{
\item{size}{circle size}

\item{cex}{character expansion factor relative to current par("cex").
Used for text, and provides the default for pt.cex.}

\item{pieCol}{colors}

\item{...}{pass to plot function}

\item{lenendPOS}{one of "bottomright", "bottom", "bottomleft", "left",
"topleft", "top", "topright", "right" and "center".}
}
\description{
plotHapNet
}
