% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapVisualization.R
\name{plotHapTable}
\alias{plotHapTable}
\title{plotHapTable}
\usage{
plotHapTable(hapResult,
hapPrefix = "H",
geneID = "",
title.color = "grey90")
}
\arguments{
\item{hapResult}{hapResult}

\item{hapPrefix}{hapPrefix}

\item{geneID}{geneID}

\item{title.color}{title.color}
}
\value{
ggplot2 object.
}
\description{
plot you hap result as a table
}
\examples{

data("quickHap_test")
hap <- get_hap(vcf,hyb_remove = TRUE, na.drop = TRUE)
hapResult <- hap_result(hap)
plotHapTable(hapResult)
}
