% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqs2hap.R
\name{seqs2hap}
\alias{seqs2hap}
\alias{allignSeqs}
\alias{trimSeqs}
\title{generate hap results from seqs}
\usage{
seqs2hap(seqs,
         Ref = names(seqs)[1],
         hyb_remove = TRUE, na.drop = TRUE,
         maxGapsPerSeq = 0.25,
         hapPrefix = "H", ...)

allignSeqs(seqs, ...)

trimSeqs(seqs,
         minFlankFraction = 0.1)
}
\arguments{
\item{seqs}{DNAStringSet or DNAMultipleAlignment}

\item{Ref}{default as name of first seq}

\item{hyb_remove}{whether remove accessions contains hybrid site or not}

\item{na.drop}{whether Drop accessions contains "N"
Default as \code{TRUE}.}

\item{maxGapsPerSeq}{A value in \verb{[0, 1]} that indicates the maximum
fraction of gaps allowed in each seq after alignment. (default is 0.25)
Seqs with gap percent exceed that will be dropped.}

\item{hapPrefix}{Prefix of hap names, default as "H".}

\item{...}{parameters will pass to muscle::muscle() for DNA multi alignment}

\item{minFlankFraction}{A value in \verb{[0, 1]} that indicates the minimum
fraction needed to call a gap in the consensus string (default is 0.1).}
}
\description{
generate hap results from seqs

allign Seqs
}
\examples{

\dontrun{
seqs <- allignSeqs(seqs)
}
\dontrun{
trimSeqs(seqs,
         minFlankFraction = 0.1)
}
}
\seealso{
\code{\link[muscle:muscle]{muscle::muscle()}}
}
