% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setATGas0.R
\name{SetATGas0}
\alias{SetATGas0}
\alias{gffSetATGas0}
\alias{hapSetATGas0}
\title{Set position of ATG as zero}
\usage{
gffSetATGas0(gff = gff, hap = hap,
             geneID = geneID,
             Chr = Chr, POS = c(start, end))

hapSetATGas0(gff = gff, hap = hap,
             geneID = geneID,
             Chr = Chr, POS = c(start, end))
}
\arguments{
\item{gff}{gff}

\item{hap}{hap results}

\item{geneID}{geneID}

\item{Chr}{Chrom name}

\item{POS}{vector defined by \code{start} and \code{end} position}
}
\value{
\code{gffSetATGas0}: filtered gff with position of ATG was as zero

\code{hapSetATGas0}: hap results only position of ATG was set as zero
}
\description{
Filter hap result and gff annotation according to provided information.
And then set position of ATG as zero in hap result and gff annotation.
The upstream was negative while the gene range and downstream was positive.

\strong{Notice:} the position of "ATG" after modified was 0, 1 and 2 separately.
The site in hap result exceed the selected range will be \strong{dropped}.
}
\examples{
# load example dataset
data("quickHap_test")

# generate hap results
hap <- vcf2hap(vcf)

# set position of ATG as zero in gff
newgff <- gffSetATGas0(gff = gff, hap = hap,
                       geneID = "test1G0387",
                       Chr = "scaffold_1",
                       POS = c(4300, 7910))

# set position of ATG as zero in hap results
newhap <- hapSetATGas0(gff = gff, hap = hap,
                       geneID = "test1G0387",
                       Chr = "scaffold_1",
                       POS = c(4300, 7910))

# visualization mutations on gene model with newgff and newhap
plotGeneStructure(gff = newgff, hapResult = newhap)

}
\seealso{
\code{\link[geneHapR:plotGeneStructure]{plotGeneStructure()}}
}
