% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionalSites.R
\name{siteEff}
\alias{siteEff}
\title{Find site functional effective}
\usage{
siteEff(hap, pheno, phenoNames)
}
\arguments{
\item{hap}{object of "hapResult" class}

\item{pheno}{phenotype data, with column names as pheno name
and row name as accessions.}

\item{phenoNames}{pheno names used for analysis, if missing,
will use all pheno names in \code{pheno}}
}
\value{
matrix, with column name are pheno names and
row name are site position
}
\description{
Find site functional effective
}
\examples{
\donttest{
data("geneHapR_test")

# calculate site functional effect
siteEFF <- siteEff(hapResult, pheno, names(pheno))
plotEff(siteEFF, gff = gff, Chr = "scaffold_1")
}
}
