% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2hap.R
\name{vcf2hap}
\alias{vcf2hap}
\title{generat haps from vcf}
\usage{
vcf2hap(vcf,
        hapPrefix = "H",
        filter_Chr = FALSE,
        Chr = Chr,
        filter_POS = FALSE,
        startPOS = startPOS,
        endPOS = endPOS,
        hyb_remove = TRUE,
        na.drop = TRUE, file = file)
}
\arguments{
\item{vcf}{vcfR object imported by \code{import_vcf}.}

\item{hapPrefix}{Prefix of hap names, default as "H".}

\item{filter_Chr}{logical, whether filter vcf by chromosome or not, default
as \code{FALSE}. If set as \code{TRUE}, the \code{Chr} is needed.}

\item{Chr}{Chromosome name, needed when \code{filter_Chr} was set as \code{TRUE}.}

\item{filter_POS}{logical, whether filter vcf by position or not. Default
as \code{FALSE}. If set as \code{TRUE}, \code{startPOS} and \code{endPOS} are needed.}

\item{startPOS, endPOS}{Start and end position, needed when \code{filter_POS} was
set as \code{TRUE}. In addition, \code{startPOS} must less than \code{endPOS}.}

\item{hyb_remove}{whether remove accessions contains hybrid site or not.
Default as \code{TRUE}.}

\item{na.drop}{whether drop accessions contains unknown allele site or not.
Default as \code{TRUE}.}

\item{file}{file path where to save the hap result. If missing, nothing will
be saved to disk.}
}
\value{
\code{haptypes} object, first four rows are meta information:
\code{CHR}, \code{POS}, \code{INFO}, \code{ALLELE.}
\code{Hap} names were located in the first column, \code{Accessions} were placed at
the last column.
}
\description{
Generate hap result from \code{vcfR} object.
A simple filter by position was provided in this function,
however it's prefer to filter vcf (\code{vcfR} object) through
\code{\link[geneHapR:filter_vcf]{filter_vcf()}}.
}
\details{
The class of returned object actually is \code{data.frame} with few
added \code{attributes} generated during processing and the data.frame itself,
such as \code{options}, \code{AccRemoved}, \code{AccRemain}, \code{AccAll}, \code{hap2acc}, \code{freq}.
Though, only a summary of the \code{attributes} were displayed, the full
information could be showed with \code{attr(x, "options")} or \code{attributes(x)}.
}
\note{
inhet test
description  generate hap format from vcf.
}
\examples{
\dontrun{
# load example data
data("quickHap_test")

# generate hap result from vcfR object
hap <- vcf2hap(vcf, hyb_remove = TRUE, na.drop =TRUE)

# generate hap result from DNAStringSet object
hap <- seqs2hap(seqs, Ref = names(seqs)[1], needAlign = TRUE, needTrim = TRUE)

# summary hap result for plot hap table
hapResult <- hap_result(hap)

# plot hap table
plotHapTable(hapResult)

# user could change/add themes, title, ...
p <- plotHapTable(hapResult)
p + ggplot2::theme_classic()
p + ggplot2::labs(title = "new title")
}
}
\seealso{
extract genotype from vcf:
\code{\link[vcfR:extract_gt_tidy]{vcfR::extract_gt_tidy()}},
import vcf files:
\code{\link[geneHapR:import_vcf]{import_vcf()}} (preferred) and
\code{\link[vcfR:read.vcfR]{vcfR::read.vcfR()}},
filter vcf according \strong{position} and \strong{annotations}:
\code{\link[geneHapR:filter_vcf]{filter_vcf()}}
}
\author{
Zhangrenl
}
